Added SmartBox Temperature option

+++ aavs_utils.py
@@ -294,7 +294,7 @@ def get_sbtemp(start=0, stop=2585699200):
         for d in data:
             if start <= int(d.split()[0]) <= stop:
                 tempi += [int(d.split()[0])]
-                dati += [int(d.split()[3])]
+                dati += [float(d.split()[3])]
     else:
         print "Unable to find the Data Logger file (/storage/monitoring/data_logger/AAVS2_Data_Logger.txt)"
     return tempi, dati
